Separate helper functions to another library file.
